Fix bugs in unmap
First bug is in state->refcount where it was incremented after
the call to async_handle_cmd. This can cause issues if
handle_unmap_cbk gets called first before the return to increment
the value of state->refcount.

Second bug is not returning the right error code when doing a
unmap_state_alloc.

Signed-off-by: Bryant G. Ly <bly@catalogicsoftware.com>